pyproject.toml
[project]
name = "aer-api-proxy"
version = "1.0.0"
requires-python = "==3.13.*"
dependencies = [
    "httpx",
    "fastapi[standard]",
    "playwright",
    "onnxruntime",
    "transformers",
    "tokenizers",
    "pillow",
    "llama-cpp-python @ git+https://github.com/JamePeng/llama-cpp-python@626a7ed33ae917d0517d7088c14a800d8f029e8f",
]
